refactor: dedup cabi fill loops into shared helpers
- _internal: add fill_with / fill_chunk / chunk_seed32; replace 55
  scalar fill loops, 4 unrolled SIMD loops, 7 chunk_seed copies
- extern C signatures untouched (csbindgen parses source without
  macro expansion); generated RngNative.cs verified byte-identical
- test: replace hardcoded sampler expectations (stale since randi/
  randf rework in 0901b48) with statistical assertions
- clippy: 88 -> 6 warnings (Safety docs, copy_from_slice, let
  chains); allow too_many_arguments / needless_range_loop (C ABI,
  SIMD kernels)

+++ src/rng32/xoshiro.rs
@@ -272,7 +272,7 @@ impl Xoshiro128Ppx16 {
     /// Must only be called on a CPU that supports AVX-512F.
     #[target_feature(enable = "avx512f")]
     pub unsafe fn randi(&mut self, min: i32, max: i32) -> [i32; 16] {
-        let v_range = _mm512_set1_epi64((max as i64 - min as i64 + 1) as i64);
+        let v_range = _mm512_set1_epi64(max as i64 - min as i64 + 1  );
         let v_min = _mm512_set1_epi32(min);
         unsafe { std::mem::transmute(self.randi_vec(v_range, v_min)) }
     }
@@ -445,7 +445,7 @@ impl Xoshiro128Ssx16 {
     /// Must only be called on a CPU that supports AVX-512F.
     #[target_feature(enable = "avx512f")]
     pub unsafe fn randi(&mut self, min: i32, max: i32) -> [i32; 16] {
-        let v_range = _mm512_set1_epi64((max as i64 - min as i64 + 1) as i64);
+        let v_range = _mm512_set1_epi64(max as i64 - min as i64 + 1  );
         let v_min = _mm512_set1_epi32(min);
         unsafe { std::mem::transmute(self.randi_vec(v_range, v_min)) }
     }
